replace faulty test case - [wip] not passing
